requirements.txt
azure-common
azure-core
azure-identity
azure-mgmt-compute
azure-mgmt-core
azure-mgmt-resource
certifi
cffi
charset-normalizer
cryptography
idna
isodate
msal
msal-extensions
msrest
oauthlib
portalocker
pycparser
PyJWT
requests
requests-oauthlib
six
urllib3
